+++ src/KardFinancial.Test/Unit/MockServer/Organizations/Placements/UpdateTest.cs
@@ -30,7 +30,6 @@ public class UpdateTest : BaseMockServerTest
               "id": "id",
               "attributes": {
                 "name": "name",
-                "status": "ACTIVE",
                 "organizationId": "organizationId",
                 "availableSlots": 1,
                 "contentStrategyId": "contentStrategyId"